ci: Split up tasks
There's no need to lint files or update codecov on multiple nodejs runtimes, so currently we have split the jobs so that we lint on latest node, then run tests on all supported node versions (except latest), and thne run tests and update codecov on latest. As a sidebonus, we now also check that the files are prettier-ed as part of the lint step.